requirements.txt
fastapi==0.112.0
uvicorn==0.30.1
sqlalchemy==2.0.32
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.36.0
